+++ src/response.js
@@ -85,8 +85,7 @@ export default class Response extends Body {
 			headers: this.headers,
 			ok: this.ok,
 			redirected: this.redirected,
-			size: this.size,
-			timeout: this.timeout
+			size: this.size
 		});
 	}